requirements/requirements.txt
neon-utils[network]~=1.0,>=1.0.3a6
ovos_utils~=0.0,>=0.0.25a10
tornado~=6.0
mycroft-messagebus-client~=0.9,!=0.9.2,!=0.9.3
ovos-core~=0.0.4
ovos-config~=0.0,>=0.0.3
